requirements.txt
streamlit
requests
pillow>=10.0.0
huggingface_hub
python-dotenv
